config: make config more flexible
Can now be passed as a `CONFIG` environment variable in a YAML format.
