+++ docker-compose.yaml
@@ -6,7 +6,8 @@ services:
     network_mode: host
     restart: always
     environment:
-      USER_ID: "@sfu:shadowfax"
-      HOMESERVER_URL: "http://localhost:8008"
-      ACCESS_TOKEN: "..."
-      TIMEOUT: 30
+      CONFIG: |
+        homeserverurl: "http://localhost:8008"
+        userid: "@sfu:shadowfax"
+        accesstoken: "..."
+        timeout: 30
